+++ .gitattributes
@@ -20,7 +20,7 @@
 /LICENSE linguist-generated
 /package.json linguist-generated
 /src/app/download/download-function.ts linguist-generated
-/src/app/formOverview/get-form-overview/getFormOverview-function.ts linguist-generated
+/src/app/formOverview/getFormOverview/getFormOverview-function.ts linguist-generated
 /src/app/listSubmissions/listSubmissions-function.ts linguist-generated
 /src/app/submission/submission-function.ts linguist-generated
 /src/migrations/migration-2024-02-06-enrich-table-function.ts linguist-generated
